+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Enhancements
 * <New feature description> (PR [#????](https://github.com/realm/realm-core/pull/????))
 * Move the implementation of the Audit API to the open-source repo and update it to work with MongoDB Realm.
+* Expose delete app user for C API. ([#5490](https://github.com/realm/realm-core/issues/5490))
+* Expose an API to get the app from user in the C API. ([#5478](https://github.com/realm/realm-core/issues/5478))
 
 ### Fixed
 * <How do the end-user experience this issue? what was the impact?> ([#????](https://github.com/realm/realm-core/issues/????), since v?.?.?)